Made the place to insert the element configurable
Added two new options for defining where to the tipsy tooltip should be added in the dom tree.

One option, insertTo, defines which element it should be inserted in relation to and the other new option, insertMethod, defines how to add the tooltip.

Code has been added to adjust the position of the tooltip according to the position of the tooltip elements offsetParent - this makes sure that it's always positioned in the right place no matter where in the dom tree it is inserted.

An advantage of adding the tooltip deeper into the dom tree is that it may move together with that content if eg. the position of a common container was to be changed - which eg. happens on a centered site whenever the browser is resized.